наконец то закончил

+++ src/main/java/com.donntu.lab7/db/repos/DepartureRepository.java
@@ -3,5 +3,5 @@ package com.donntu.lab7.db.repos;
 import com.donntu.lab7.db.entities.DepartureDate;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DepartureRepository extends CrudRepository<DepartureDate, Integer> {
+public interface DepartureRepository extends CrudRepository<DepartureDate, Long> {
 }
